Small change to avoid integer overflow for "Goldilocks" problems.

+++ packages/zoltan2/src/problems/Zoltan2_PartitioningSolution.hpp
@@ -1273,7 +1273,7 @@ template <typename Adapter>
         partId_t *tmpPart = new partId_t [newLen];
         env_->localMemoryAssertion(__FILE__, __LINE__, newLen, tmpPart);
     
-        int next = 0;
+        size_t next = 0;
         for (lno_t i=0; i < newLen; i++){
           tmpGno[i] = inBuf[next++];
           tmpPart[i] = inBuf[next++];
